* Huge change in the Color Conversion System : introduce profile, which means nodes contains the color model id, the depth id _and_ the profile name, which means there is a node for a KoColorSpace/KoColorProfile couple. * make icc color space be less of an exception, introduce KoColorSpaceEngine, this class allows to create color conversion transformation for a "class" of color space, for instance, for ICC to ICC conversion is allways done the same way, this allow to reduce the number of connection in the CCS graph (reducing memory consuption and cpu time when looking for the best path)
(disable compilation of xyz_float_hdr, if someone wants to port it to the change, feel free)

svn path=/trunk/koffice/; revision=789697

+++ libs/pigment/colorspaces/KoRgbU8ColorSpace.cc
@@ -124,5 +124,5 @@ void KoRgbU8ColorSpace::colorFromXML( quint8* pixel, const QDomElement& elt) con
 quint8 KoRgbU8ColorSpace::intensity8(const quint8 * src) const
 {
     const RgbU8Traits::Pixel* p = reinterpret_cast<const RgbU8Traits::Pixel*>( src );
-    return p->red * 0.30 + p->green * 0.59 + p->blue * 0.11;
+    return (quint8)(p->red * 0.30 + p->green * 0.59 + p->blue * 0.11);
 }
